Switching to GitHub Actions (#77)
Switching to GitHub Actions
=====================

This PR switches ToMCAT over to testing using GitHub Actions instead of Travis CI. 

Major improvements
--------------------

- This reduces the automated build and testing time from ~16 minutes to ~7 minutes, a significant improvement.
- It also enables us to test the Java portions of the codebase, since it's not such a hassle to use Java 1.8
- The PR also implements testing the run_session.sh script, which is becoming more and more important. It does so using Xvfb.


Minor changes
---------------

- The `launch_minecraft_in_background.py` from Malmo has been removed
- The required version of Java (1.8)  is now enforced by `external/malmo/CMakeLists.txt`
- Ran some scripts through ShellCheck, fixed some common potential issues.

+++ tools/run_session.sh
@@ -2,34 +2,42 @@
 
 set -u
 
+declare -x GITHUB_ACTIONS
+
 # this function returns the date and time in the current timezone.
 timestamp() {
-    date "+%Y_%m_%d_%H_%M_%S"
+  date "+%Y_%m_%d_%H_%M_%S"
 }
 
-mission_one_time=600
-do_tutorial=1
+if [[ -n ${GITHUB_ACTIONS} ]]; then
+  time_limit=5
+  do_tutorial=0
+else
+  time_limit=600
+  do_tutorial=1
+fi
 do_invasion=1
 
 # Set the TOMCAT environment variable, assuming that the directory structure
 # mirrors that of the git repository.
-export TOMCAT="$(cd "$(dirname "${BASH_SOURCE[0]}")/../" >/dev/null 2>&1 && pwd)"
+TOMCAT="$(cd "$(dirname "${BASH_SOURCE[0]}")/../" >/dev/null 2>&1 && pwd)"
+export TOMCAT
+
 export TOMCAT_TMP_DIR="/tmp/$USER/tomcat"
 mkdir -p "${TOMCAT_TMP_DIR}"
 
-${TOMCAT}/tools/check_minecraft.sh
-if [[ $? -ne 0 ]]; then exit 1; fi
+if ! "${TOMCAT}"/tools/check_minecraft.sh; then exit 1; fi
 
 export tutorial_mission_log="${TOMCAT_TMP_DIR}/tutorial_mission.log"
 export zombie_invasion_log="${TOMCAT_TMP_DIR}/zombie_invasion.log"
 
 export num_tries=2
 
-if [[ ${do_tutorial} -eq 1 ]]; then
-    ${TOMCAT}/tools/run_tutorial
+if [[ ${do_tutorial} -eq 1 ]]; then 
+  "${TOMCAT}"/tools/run_tutorial
 fi
 
-rm -f ${TOMCAT}/external/malmo/Minecraft/run/saves/discrete_events/discrete_events.json
+rm -f "${TOMCAT}"/external/malmo/Minecraft/run/saves/discrete_events/discrete_events.json
 
 if [[ ${do_invasion} -eq 1 ]]; then
     echo " "
@@ -40,7 +48,7 @@ if [[ ${do_invasion} -eq 1 ]]; then
 
     framerate_option=""
     if [[ "$OSTYPE" == "darwin"* ]]; then
-       osascript ${TOMCAT}/tools/activate_minecraft_window.scpt
+        osascript "${TOMCAT}"/tools/activate_minecraft_window.scpt
 
         # On macOS, we choose the avfoundation format.
         ffmpeg_fmt=avfoundation
@@ -59,36 +67,47 @@ if [[ ${do_invasion} -eq 1 ]]; then
     fi
 
     # Creating an output directory for this session.
-    output_dir=${TOMCAT}/data/participant_data/session_$(timestamp)
-    mkdir -p ${output_dir}
+    output_dir="${TOMCAT}"/data/participant_data/session_$(timestamp)
+    mkdir -p "${output_dir}"
     ffmpeg_common_invocation="ffmpeg -nostdin -f $ffmpeg_fmt"
 
-    # Recording video of player's face
-    ${ffmpeg_common_invocation} ${framerate_option} \
-        -i "0:" ${output_dir}/webcam_video.mpg &>/dev/null &
-    webcam_recording_pid=$!
+    if [[ -z $GITHUB_ACTIONS ]]; then
+        # Recording video of player's face
+        "${ffmpeg_common_invocation}" "${framerate_option}" \
+            -i "0:" "${output_dir}"/webcam_video.mpg &>/dev/null &
+        webcam_recording_pid=$!
 
-    # Recording player audio
-    ${ffmpeg_common_invocation} \
-        -i ":0" ${output_dir}/player_audio.wav &>/dev/null &
-    audio_recording_pid=$!
+        # Recording player audio
+        "${ffmpeg_common_invocation}" \
+            -i ":0" "${output_dir}"/player_audio.wav &>/dev/null &
+        audio_recording_pid=$!
+    fi
 
     # Recording game screen.
     if [[ "$OSTYPE" == "darwin"* ]]; then
-        ${ffmpeg_common_invocation} \
-            -i "1:" -r 30 ${output_dir}/screen_video.mpg &> /dev/null &
+        "${ffmpeg_common_invocation}" \
+            -i "1:" -r 30 "${output_dir}"/screen_video.mpg &> /dev/null &
         screen_recording_pid=$!
     fi
 
     while [ $try -lt $num_tries ]; do
-        ${TOMCAT}/build/bin/runExperiment \
+        if [[ -n "$GITHUB_ACTIONS" ]]; then
+            "${TOMCAT}"/build/bin/runExperiment \
+            --mission external/malmo/sample_missions/default_flat_1.xml\
+            --time_limit ${time_limit} \
+            --record_path "${output_dir}"/malmo_data.tgz \
+            &>"${zombie_invasion_log}" &
+            bg_pid=$!
+        else
+            "${TOMCAT}"/build/bin/runExperiment \
             --mission 1 \
-            --time_limit ${mission_one_time} \
-            --record_path "${output_dir}/malmo_data.tgz" \
-            &>${zombie_invasion_log} &
-        bg_pid=$!
+            --time_limit ${time_limit} \
+            --record_path "${output_dir}"/malmo_data.tgz \
+            &>"${zombie_invasion_log}" &
+            bg_pid=$!
+        fi
         echo "Running: ${TOMCAT}/build/bin/runExperiment --mission 1"
-        echo "    --time_limit ${mission_one_time}"
+        echo "    --time_limit ${time_limit}"
         echo "    --record_path ${output_dir}/malmo_data.tgz"
         echo "Process corresponding to ./bin/runExperiment is $bg_pid"
         echo "... waiting for it to complete."
@@ -97,7 +116,7 @@ if [[ ${do_invasion} -eq 1 ]]; then
 
         if [[ ${zombie_invasion_status} -eq 0 ]]; then
             zombie_invasion_status=$(grep -c "Error starting mission" \
-                ${zombie_invasion_log})
+            "${zombie_invasion_log}")
         fi
 
         if [[ ${zombie_invasion_status} -eq 0 ]]; then
@@ -107,19 +126,23 @@ if [[ ${do_invasion} -eq 1 ]]; then
             break
         fi
 
-        let try+=1
+        (( try++ ))
 
         if [[ $try -lt $num_tries ]]; then
             echo "Zombie invasion mission ended with failure status."
-            echo "Killing all Minecraft and Malmo processes that can be found and trying again."
-            ${TOMCAT}/tools/kill_minecraft.sh
-            ${TOMCAT}/tools/check_minecraft.sh
+            echo "Killing all Minecraft and Malmo processes that can be found"
+            echo "and trying again."
+
+            "${TOMCAT}"/tools/kill_minecraft.sh
+            "${TOMCAT}"/tools/check_minecraft.sh
         fi
     done
 fi
 
-kill -2 $webcam_recording_pid
-kill -2 $audio_recording_pid
+if [[ -z $GITHUB_ACTIONS ]]; then
+    kill -2 $webcam_recording_pid
+    kill -2 $audio_recording_pid
+fi
 
 # For now, screen recording works only on macOS. Need to extend it to Linux as
 # well.
@@ -127,7 +150,13 @@ if [[ "$OSTYPE" == "darwin"* ]]; then
     kill -2 $screen_recording_pid
 fi
 
-mv ${TOMCAT}/external/malmo/Minecraft/run/saves/discrete_events/discrete_events.json ${output_dir}/discrete_events.json 
+# When performing continuous integration with Github Actions, we run a test
+# mission that doesn't produce the JSON file corresponding to discrete actions,
+# So we only move it in a non-Github Actions environment.
+if [[ -z $GITHUB_ACTIONS ]]; then
+  mv "${TOMCAT}"/external/malmo/Minecraft/run/saves/discrete_events/discrete_events.json\
+     "${output_dir}"/discrete_events.json
+fi
 
-echo "Finished running all sessions in ${TOMCAT}."
+echo "Finished running all sessions in ${TOMCAT}".
 exit 0
